% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTrack-class.R
\docType{class}
\name{CustomTrack-class}
\alias{CustomTrack-class}
\alias{initialize,CustomTrack-method}
\alias{CustomTrack}
\alias{drawGD,CustomTrack-method}
\alias{show,CustomTrack-method}
\title{CustomTrack class and methods}
\usage{
\S4method{initialize}{CustomTrack}(.Object, plottingFunction, variables, ...)

CustomTrack(
  plottingFunction = function(GdObject, prepare = FALSE, ...) {
 },
  variables = list(),
  name = "CustomTrack",
  ...
)

\S4method{drawGD}{CustomTrack}(GdObject, minBase, maxBase, prepare = FALSE, ...)

\S4method{show}{CustomTrack}(object)
}
\arguments{
\item{.Object}{The object skeleton passed on by \code{new()} during class
instantiation, to be filled in by the \code{initialize} method.}

\item{plottingFunction}{A user-defined function to be executed once the track
coordinates have been properly set up. The function needs to accept two
mandatory arguments: \code{\link[=GdObject-class]{GdObject}}, the \code{CustomTrack} object
to be plotted, and \code{prepare}, a logical flag indicating whether the function
has been called in preparation mode or in drawing mode. It also needs to
return the input \code{\link[=GdObject-class]{GdObject}}, potentially with
modifications.}

\item{variables}{A list of additional variables for the user-defined
plotting function.}

\item{...}{Additional items which will all be interpreted as further
display parameters. See \code{\link{settings}} and the "Display Parameters"
section below for details.}

\item{name}{Character scalar of the track's name.}

\item{GdObject}{Object of class \code{\link[=GdObject-class]{GdObject}}.}

\item{minBase}{Numeric scalar, the start coordinate of the plotting range.}

\item{maxBase}{Numeric scalar, the end coordinate of the plotting range.}

\item{prepare}{\code{logical}. Run the drawing method in preparation rather than
in production mode, i.e., only compute the track's layout without rendering
anything to the device.}

\item{object}{Object of class \code{\link[=CustomTrack-class]{CustomTrack}}.}
}
\value{
The return value of the constructor function is a new object of class
\code{CustomTrack}.
}
\description{
A fully customizable track object to be populated via a user-defined
plotting function.
}
\details{
A track to allow for any sort of plotting, with the currently displayed
genomic location set. Essentially this acts as a simple callback into the
\code{Gviz} plotting machinery after all the track panels and coordinates
have been set up. It is entirely up to the user what to plot in the track,
or even to use the predefined coordinate system. The only prerequisite is
that all plotting operations need to utilize Grid graphics.
}
\section{Functions}{
\itemize{
\item \code{initialize(CustomTrack)}: Initialize the \code{plottingFunction} and
\code{variables} slots before deferring to the \code{\link[=GdObject-class]{GdObject}}
initializer for the remaining slots.

\item \code{CustomTrack()}: Objects can be created using the constructor
function.

\item \code{drawGD(CustomTrack)}: plot the object to a graphics device.
The return value of this method is the input object, potentially updated
during the plotting operation. Internally, there are two modes in which the
method can be called. Either in 'prepare' mode, in which case no plotting is
done but the object is preprocessed based on the available space, or in
'plotting' mode, in which case the actual graphical output is created.
Since subsetting of the object can be potentially costly, this can be
switched off in case subsetting has already been performed before or
is not necessary.

\item \code{show(CustomTrack)}: Show method.

}}
\examples{
## Object construction:

## An empty object
CustomTrack()
}
\seealso{
\link[=DisplayPars-class]{DisplayPars}

\link[=GdObject-class]{GdObject}

\code{\link[GenomicRanges:GRanges-class]{GRanges}}

\link[=HighlightTrack-class]{HighlightTrack}

\link[=ImageMap-class]{ImageMap}

\code{\link[IRanges:IRanges-class]{IRanges}}

\link[=RangeTrack-class]{RangeTrack}

\link[=DataTrack-class]{DataTrack}

\link{collapsing}

\link{grouping}

\code{\link[lattice:panel.grid]{panel.grid}}

\link{plotTracks}

\link{settings}
}
\author{
Florian Hahne
}
