% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gviz-defunct.R
\name{Gviz-defunct}
\alias{Gviz-defunct}
\title{Defunct functions in package \code{Gviz}}
\value{
No return value, called for documentation purposes only.
}
\description{
These functions are defunct and no longer available.
\subsection{Defunct functions are:}{

(none)
}
}
