% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTrack-class.R
\docType{class}
\name{ReferenceDataTrack-class}
\alias{ReferenceDataTrack-class}
\title{The file-based version of the \code{DataTrack} class}
\description{
This will mainly provide a means to dispatch to a special \code{subset} method
which should stream the necessary data from disk. Users typically do not
have to deal with this distinction directly and can rely on the
\code{DataTrack} constructor to make the right choice.
}
\keyword{internal}
