% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneRegionTrack-class.R
\docType{class}
\name{ReferenceGeneRegionTrack-class}
\alias{ReferenceGeneRegionTrack-class}
\title{The file-based version of the \code{GeneRegionTrack} class}
\description{
This will mainly provide a means to dispatch to a special \code{subset} method
which should stream the necessary data from disk. Users typically do not
have to deal with this distinction directly and can rely on the
\code{GeneRegionTrack} constructor to make the right choice.
}
\keyword{internal}
