% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceTrack-class.R
\docType{class}
\name{ReferenceSequenceTrack-class}
\alias{ReferenceSequenceTrack-class}
\alias{initialize,ReferenceSequenceTrack-method}
\title{The file-based version of the \code{SequenceTrack} class}
\usage{
\S4method{initialize}{ReferenceSequenceTrack}(.Object, stream, reference, ...)
}
\description{
This will mainly provide a means to dispatch to a special \code{subseq} method
which should stream the necessary data from disk. Users typically do not
have to deal with this distinction directly and can rely on the
\code{SequenceTrack} constructor to make the right choice.
}
\section{Functions}{
\itemize{
\item \code{initialize(ReferenceSequenceTrack)}: Initialize the
\code{\link[=ReferenceTrack-class]{ReferenceTrack}} slots (\code{stream}, \code{reference}) before
deferring to the
\code{\link[=SequenceDNAStringSetTrack-class]{SequenceDNAStringSetTrack}} initializer
for the remaining slots.

}}
\keyword{internal}
