% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceTrack-class.R
\docType{class}
\name{SequenceRNAStringSetTrack-class}
\alias{SequenceRNAStringSetTrack-class}
\alias{initialize,SequenceRNAStringSetTrack-method}
\title{The \code{RNAStringSet}-based version of the \code{SequenceTrack} class}
\usage{
\S4method{initialize}{SequenceRNAStringSetTrack}(.Object, sequence, ...)
}
\arguments{
\item{.Object}{The object skeleton passed on by \code{new()} during class
instantiation, to be filled in by the \code{initialize} method.}

\item{sequence}{An \code{\link[Biostrings:RNAStringSet-class]{RNAStringSet}} object
with one RNA sequence per chromosome. The names have to match the chromosome
identifiers.}

\item{...}{Additional arguments.}
}
\value{
An initialized object of class
\code{\link[=SequenceRNAStringSetTrack-class]{SequenceRNAStringSetTrack}}.
}
\description{
The \code{RNAStringSet}-based version of the \code{SequenceTrack} class
}
\section{Functions}{
\itemize{
\item \code{initialize(SequenceRNAStringSetTrack)}: Initialize the \code{sequence} slot
(defaulting to an empty \code{\link[Biostrings:RNAStringSet-class]{RNAStringSet}} if
none is supplied) before deferring to the \code{SequenceTrack} initializer for the
remaining slots.

}}
\section{Slots}{

\describe{
\item{\code{sequence}}{A \code{\link[Biostrings:RNAStringSet-class]{RNAStringSet}} object
containing the sequence data, with one element per chromosome.}
}}

