% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceTrack-class.R
\docType{class}
\name{SequenceTrack-class}
\alias{SequenceTrack-class}
\alias{initialize,SequenceTrack-method}
\alias{SequenceTrack}
\alias{RNASequenceTrack}
\alias{seqnames,SequenceTrack-method}
\alias{seqnames,SequenceBSgenomeTrack-method}
\alias{seqlevels,SequenceTrack-method}
\alias{seqlevels,SequenceBSgenomeTrack-method}
\alias{start,SequenceTrack-method}
\alias{end,SequenceTrack-method}
\alias{width,SequenceTrack-method}
\alias{length,SequenceTrack-method}
\alias{subseq,SequenceTrack-method}
\alias{subseq,ReferenceSequenceTrack-method}
\alias{chromosome,SequenceTrack-method}
\alias{chromosome<-,SequenceTrack-method}
\alias{genome,SequenceTrack-method}
\alias{consolidateTrack,SequenceTrack-method}
\alias{drawGD,SequenceTrack-method}
\alias{show,SequenceBSgenomeTrack-method}
\alias{show,SequenceDNAStringSetTrack-method}
\alias{show,SequenceRNAStringSetTrack-method}
\alias{show,ReferenceSequenceTrack-method}
\title{SequenceTrack class and methods}
\usage{
\S4method{initialize}{SequenceTrack}(.Object, chromosome, genome, ...)

SequenceTrack(
  sequence,
  chromosome,
  genome,
  name = "SequenceTrack",
  importFunction,
  stream = FALSE,
  ...
)

RNASequenceTrack(
  sequence,
  chromosome,
  genome,
  name = "SequenceTrack",
  importFunction,
  stream = FALSE,
  ...
)

\S4method{seqnames}{SequenceTrack}(x)

\S4method{seqnames}{SequenceBSgenomeTrack}(x)

\S4method{seqlevels}{SequenceTrack}(x)

\S4method{seqlevels}{SequenceBSgenomeTrack}(x)

\S4method{start}{SequenceTrack}(x)

\S4method{end}{SequenceTrack}(x)

\S4method{width}{SequenceTrack}(x)

\S4method{length}{SequenceTrack}(x)

\S4method{subseq}{SequenceTrack}(x, start = NA, end = NA, width = NA)

\S4method{subseq}{ReferenceSequenceTrack}(x, start = NA, end = NA, width = NA)

\S4method{chromosome}{SequenceTrack}(GdObject)

\S4method{chromosome}{SequenceTrack}(GdObject) <- value

\S4method{genome}{SequenceTrack}(x)

\S4method{consolidateTrack}{SequenceTrack}(GdObject, chromosome, ...)

\S4method{drawGD}{SequenceTrack}(GdObject, minBase, maxBase, prepare = FALSE, ...)

\S4method{show}{SequenceBSgenomeTrack}(object)

\S4method{show}{SequenceDNAStringSetTrack}(object)

\S4method{show}{SequenceRNAStringSetTrack}(object)

\S4method{show}{ReferenceSequenceTrack}(object)
}
\arguments{
\item{.Object}{The object skeleton passed on by \code{new()} during class
instantiation, to be filled in by the \code{initialize} method.}

\item{chromosome}{The currently active chromosome, which may have to be set
for a \code{\link[=RangeTrack-class]{RangeTrack}} or a \code{SequenceTrack} object.}

\item{genome}{The genome on which the track's ranges are defined. Usually
this is a valid UCSC genome identifier, however this is not being formally
checked at this point. For a
\code{\link[=SequenceBSgenomeTrack-class]{SequenceBSgenomeTrack}} object, the genome
information is extracted from the input
\code{\link[BSgenome:BSgenome-class]{BSgenome}} package. For a
\code{\link[Biostrings:DNAStringSet-class]{DNAStringSet}} it has to be provided, or
the constructor will fall back to the default value of \code{NA}.}

\item{...}{Additional items which will all be interpreted as further
display parameters. See \code{\link{settings}} and the "Display Parameters"
section below for details.}

\item{sequence}{A meta argument to handle the different input types, making
the construction of a \code{SequenceTrack} as flexible as possible.

The different input options for \code{sequence} are:
\itemize{
\item \strong{An object of class \code{\link[Biostrings:DNAStringSet-class]{DNAStringSet}}}:
the individual \code{DNAString}s are considered to be the different
chromosome sequences.
\item \strong{An object of class \code{\link[BSgenome:BSgenome-class]{BSgenome}}}: the \code{Gviz}
package tries to follow the \code{\link[BSgenome:BSgenome-class]{BSgenome}}
philosophy in that the respective chromosome sequences are only realized
once they are first accessed.
\item \strong{A \code{character} scalar}: the value of the \code{sequence} argument is
considered to be a file path to an annotation file on disk. A range of
file types are supported by the \code{Gviz} package as identified by the file
extension. See the \code{importFunction} documentation below for further
details.
}}

\item{name}{Character scalar of the track's name used in the title panel
when plotting.}

\item{importFunction}{A user-defined function to be used to import the
sequence data from a file. This only applies when the \code{sequence} argument is
a character string with the path to the input data file. The function needs
to accept an argument \code{file} containing the file path and has to return a
proper \code{\link[Biostrings:DNAStringSet-class]{DNAStringSet}} object with the
sequence information per chromosome. A set of default import functions is
already implemented in the package for a number of different file types, and
one of these defaults will be picked automatically based on the extension of
the input file name. If the extension cannot be mapped to any of the existing
import functions, an error is raised asking for a user-defined import
function. Currently the following file types can be imported with the default
functions: \code{fa/fasta} and \verb{2bit}.

Both file types support indexing by genomic coordinates, and it makes sense
to only load the part of the file that is needed for plotting. To this end,
the \code{Gviz} package defines the derived
\code{\link[=ReferenceSequenceTrack-class]{ReferenceSequenceTrack}} class, which
supports streaming data from the file system. The user typically does not
have to deal with this distinction but may rely on the constructor function
to make the right choice as long as the default import functions are used.
However, once a user-defined import function has been provided and if this
function adds support for indexed files, you will have to make the
constructor aware of this fact by setting the \code{stream} argument to \code{TRUE}.
Please note that in this case the import function needs to accept a second
mandatory argument \code{selection}, which is a
\code{\link[GenomicRanges:GRanges-class]{GRanges}} object containing the dimensions of
the plotted genomic range. As before, the function has to return an
appropriate \code{\link[Biostrings:DNAStringSet-class]{DNAStringSet}} object.}

\item{stream}{A logical flag indicating that the user-provided import
function can deal with indexed files and knows how to process the additional
\code{selection} argument when accessing the data on disk. This causes the
constructor to return a
\code{\link[=ReferenceSequenceTrack-class]{ReferenceSequenceTrack}} object which will
grab the necessary data on the fly during each plotting operation.}

\item{x}{A valid track object class name, or the object itself, in which
case the class is derived directly from it.}

\item{start, end, width}{Integer scalars defining the requested subsequence.
Exactly two of the three must be provided; the third is derived from them.}

\item{GdObject}{Object of class \code{\link[=GdObject-class]{GdObject}}.}

\item{value}{Value to be set.}

\item{minBase}{Numeric scalar, the start coordinate of the sequence.}

\item{maxBase}{Numeric scalar, the end coordinate of the sequence.}

\item{prepare}{\code{logical}. Run the drawing method in preparation rather than
in production mode, i.e., only compute the track's layout without rendering
anything to the device.}

\item{object}{Object inheriting from \code{\link[=SequenceTrack-class]{SequenceTrack}}.}
}
\value{
The return value of the constructor function is a new object of class
\code{\link[=SequenceDNAStringSetTrack-class]{SequenceDNAStringSetTrack}},
\code{\link[=SequenceBSgenomeTrack-class]{SequenceBSgenomeTrack}} or
\code{\link[=ReferenceSequenceTrack-class]{ReferenceSequenceTrack}}, depending on the
constructor arguments. Typically the user will not have to be troubled with
this distinction and can rely on the constructor to make the right choice.
}
\description{
A track class to represent genomic sequences. The three child classes
\code{\link[=SequenceDNAStringSetTrack-class]{SequenceDNAStringSetTrack}},
\code{\link[=SequenceRNAStringSetTrack-class]{SequenceRNAStringSetTrack}} and
\code{\link[=SequenceBSgenomeTrack-class]{SequenceBSgenomeTrack}} do most of the work,
however in practise they are of no particular relevance to the user.
}
\section{Functions}{
\itemize{
\item \code{initialize(SequenceTrack)}: Initialize the display parameter defaults and
the \code{chromosome}/\code{genome} slots before deferring to the
\code{\link[=GdObject-class]{GdObject}} initializer for the remaining slots.

\item \code{SequenceTrack()}: Constructor

\item \code{RNASequenceTrack()}: Constructor

\item \code{seqnames(SequenceTrack)}: return the names (i.e., the chromosome)
of the sequences contained in the object.

\item \code{seqnames(SequenceBSgenomeTrack)}: return the names (i.e., the chromosome)
of the sequences contained in the object.

\item \code{seqlevels(SequenceTrack)}: return the names (i.e., the chromosome)
of the sequences contained in the object. Only those with length > 0.

\item \code{seqlevels(SequenceBSgenomeTrack)}: return the names (i.e., the chromosome)
of the sequences contained in the object. Only those with length > 0.

\item \code{start(SequenceTrack)}: return the start coordinates of the track
items.

\item \code{end(SequenceTrack)}: return the end coordinates of the track
items.

\item \code{width(SequenceTrack)}: return the with of the track items in
genomic coordinates.

\item \code{length(SequenceTrack)}: return the length of the sequence for
active chromosome.

\item \code{subseq(SequenceTrack)}: extract a subsequence for the active
chromosome between \code{start} and \code{end} (or of the given \code{width}), padding
with \code{-} outside the available sequence range. Two of \code{start}, \code{end} and
\code{width} must be provided.

\item \code{subseq(ReferenceSequenceTrack)}: extract a subsequence for the active
chromosome by streaming the required region from the referenced file,
rather than from an in-memory sequence.

\item \code{chromosome(SequenceTrack)}: return the chromosome for which the track
is defined.

\item \code{chromosome(SequenceTrack) <- value}: replace the value of the track's chromosome.
This has to be a valid UCSC chromosome identifier or an integer or character
scalar that can be reasonably coerced into one.

\item \code{genome(SequenceTrack)}: Set the track's genome.
Usually this has to be a valid UCSC identifier, however this is not
formally enforced here.

\item \code{consolidateTrack(SequenceTrack)}: Determine whether there are \code{chromosome}
settings or not, and add this information.

\item \code{drawGD(SequenceTrack)}: plot the object to a graphics device.
The return value of this method is the input object, potentially updated
during the plotting operation. Internally, there are two modes in which the
method can be called. Either in 'prepare' mode, in which case no plotting is
done but the object is preprocessed based on the available space, or in
'plotting' mode, in which case the actual graphical output is created.
Since subsetting of the object can be potentially costly, this can be
switched off in case subsetting has already been performed before or
is not necessary.

\item \code{show(SequenceBSgenomeTrack)}: Show method.

\item \code{show(SequenceDNAStringSetTrack)}: Show method.

\item \code{show(SequenceRNAStringSetTrack)}: Show method.

\item \code{show(ReferenceSequenceTrack)}: Show method.

}}
\section{Objects from the class}{


Objects can be created using the constructor function \code{SequenceTrack}.
}

\examples{
## An empty object
SequenceTrack()

## Construct from DNAStringSet
library(Biostrings)
letters <- c("A", "C", "T", "G", "N")
set.seed(999)
seqs <- DNAStringSet(c(chr1 = paste(sample(letters, 100000, TRUE),
    collapse = ""
), chr2 = paste(sample(letters, 200000, TRUE), collapse = "")))
sTrack <- SequenceTrack(seqs, genome = "hg38")
sTrack

## Construct from BSgenome object
if (require(BSgenome.Hsapiens.UCSC.hg38)) {
    sTrack <- SequenceTrack(Hsapiens)
    sTrack
}


## Set active chromosome
chromosome(sTrack)
chromosome(sTrack) <- "chr2"
head(seqnames(sTrack))
\dontshow{
## For some annoying reason the postscript device does not know about
## the sans font
if (!interactive()) {
    font <- ps.options()$family
    displayPars(sTrack) <- list(fontfamily = font, fontfamily.title = font)
}
}

## Plotting
## Sequences
plotTracks(sTrack, from = 199970, to = 200000)
## Boxes
plotTracks(sTrack, from = 199800, to = 200000)
## Line
plotTracks(sTrack, from = 1, to = 200000)
## Force boxes
plotTracks(sTrack, from = 199970, to = 200000, noLetters = TRUE)
## Direction indicator
plotTracks(sTrack, from = 199970, to = 200000, add53 = TRUE)
## Sequence complement
plotTracks(sTrack,
    from = 199970, to = 200000, add53 = TRUE, complement = TRUE
)
## Colors
plotTracks(sTrack, from = 199970, to = 200000, add53 = TRUE, fontcolor = c(
    A = 1,
    C = 1, G = 1, T = 1, N = 1
))

## Track names
names(sTrack)
names(sTrack) <- "foo"

## Accessors
genome(sTrack)
genome(sTrack) <- "mm39"
length(sTrack)

## Sequence extraction
subseq(sTrack, start = 100000, width = 20)
## beyond the stored sequence range
subseq(sTrack, start = length(sTrack), width = 20)
}
\seealso{
\link[=DisplayPars-class]{DisplayPars}

\link[=GdObject-class]{GdObject}

\code{\link[GenomicRanges:GRanges-class]{GRanges}}

\link[=HighlightTrack-class]{HighlightTrack}

\link[=ImageMap-class]{ImageMap}

\code{\link[IRanges:IRanges-class]{IRanges}}

\link[=RangeTrack-class]{RangeTrack}

\link[=DataTrack-class]{DataTrack}

\link{collapsing}

\link{grouping}

\code{\link[lattice:panel.grid]{panel.grid}}

\link{plotTracks}

\link{settings}
}
\author{
Florian Hahne
}
