% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{bmTrack}
\alias{cyp2b10}
\alias{idTrack}
\alias{biomTrack}
\alias{biomTrack2}
\alias{cpgIslands}
\alias{axTrack}
\alias{conservation}
\alias{ensGenes}
\alias{denseAnnTrack}
\alias{geneModels}
\alias{iTrack}
\alias{itrack}
\alias{idxTrack}
\alias{ideoTrack}
\alias{twoGroups}
\alias{from}
\alias{gcContent}
\alias{knownGenes}
\alias{refGenes}
\alias{snpLocations}
\alias{to}
\alias{ctrack}
\alias{geneDetails}
\alias{dtHoriz}
\alias{bmt}
\title{Data sets}
\description{
Some sample data sets used for the illustrative examples and the vignette.
}
\keyword{datasets}
