\name{bmTrack}
\alias{bmTrack}
\alias{cyp2b10}
\alias{idTrack}
\alias{biomTrack}
\alias{cpgIslands}
\alias{axTrack}
\alias{conservation}
\alias{ensGenes}
\alias{denseAnnTrack}
\alias{geneModels}
\alias{ideoTrack}
\alias{itrack}
\alias{twoGroups}
\alias{from}
\alias{gcContent}
\alias{idxTrack}
\alias{knownGenes}
\alias{refGenes}
\alias{snpLocations}
\alias{to}
\alias{ctrack}
\alias{geneDetails}
\docType{data}
\title{
Data sets
}

\description{
Some sample data sets used for the illustrative examples and the vignette.
}

\keyword{datasets}





