% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{grouping}
\alias{grouping}
\title{Grouping of annotation features}
\description{
Many annotation tracks are actually composed of a number of grouped
sub-features, for instance exons in a gene model. This man page highlights
the use of grouping information to build informative annotation plots.
}
\details{
All track objects that inherit from class
\code{\link[=AnnotationTrack-class]{AnnotationTrack}} support the grouping feature. The
information is usually passed on to the constructor function (for
\code{\link[=AnnotationTrack-class]{AnnotationTrack}} via the \code{group} argument and for
\code{\link[=GeneRegionTrack-class]{GeneRegionTrack}} objects via the \code{exon} argument)
or automatically downloaded from an online annotation repository
(\code{\link[=BiomartGeneRegionTrack-class]{BiomartGeneRegionTrack}}). Group membership
is specified by a factor vector with as many items as there are annotation
items in the track (i.e., the value of \code{length(track)}). Upon plotting, the
grouped annotation features are displayed together and will not be separated
in the stacking of track items.
}
\seealso{
\link[=AnnotationTrack-class]{AnnotationTrack}

\link[=BiomartGeneRegionTrack-class]{BiomartGeneRegionTrack}

\link[=GeneRegionTrack-class]{GeneRegionTrack}
}
\author{
Florian Hahne
}
