\name{Gviz-deprecated}
\alias{Gviz-deprecated}
\title{Deprecated functions in package \sQuote{Gviz}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{Gviz} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
}
