% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gviz-deprecated.R
\name{Gviz-deprecated}
\alias{Gviz-deprecated}
\title{Deprecated functions in package \code{Gviz}}
\description{
These functions are provided for compatibility with older
versions of \code{Gviz} only, and will be defunct at the next release.
\subsection{The following functions are deprecated and will be made defunct (use the replacement indicated below):}{

(none)
}
}
