% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceTrack-class.R
\docType{class}
\name{SequenceBSgenomeTrack-class}
\alias{SequenceBSgenomeTrack-class}
\alias{initialize,SequenceBSgenomeTrack-method}
\title{The \code{BSgenome}-based version of the \code{SequenceTrack} class}
\usage{
\S4method{initialize}{SequenceBSgenomeTrack}(.Object, sequence = NULL, ...)
}
\arguments{
\item{.Object}{The object skeleton passed on by \code{new()} during class
instantiation, to be filled in by the \code{initialize} method.}

\item{sequence}{A \code{\link[BSgenome:BSgenome-class]{BSgenome}} object with the full
genomic sequence, or \code{NULL}.}

\item{...}{Additional arguments.}
}
\description{
The \code{BSgenome}-based version of the \code{SequenceTrack} class
}
\section{Functions}{
\itemize{
\item \code{initialize(SequenceBSgenomeTrack)}: Initialize the \code{sequence} and
\code{pointerCache} slots before deferring to the \code{SequenceTrack} initializer
for the remaining slots.

}}
\section{Slots}{

\describe{
\item{\code{sequence}}{A \code{\link[BSgenome:BSgenome-class]{BSgenome}} object (or \code{NULL}),
realized on demand as individual chromosome sequences are accessed.}

\item{\code{pointerCache}}{An environment holding pointers to the
\code{\link[BSgenome:BSgenome-class]{BSgenome}} sequences that have already been
realized, to prevent garbage collection.}
}}

