% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceTrack-class.R
\docType{class}
\name{SequenceDNAStringSetTrack-class}
\alias{SequenceDNAStringSetTrack-class}
\alias{initialize,SequenceDNAStringSetTrack-method}
\title{The \code{DNAStringSet}-based version of the \code{SequenceTrack} class}
\usage{
\S4method{initialize}{SequenceDNAStringSetTrack}(.Object, sequence, ...)
}
\arguments{
\item{.Object}{The object skeleton passed on by \code{new()} during class
instantiation, to be filled in by the \code{initialize} method.}

\item{sequence}{A \code{\link[Biostrings:DNAStringSet-class]{DNAStringSet}} object
with one DNA sequence per chromosome. The names have to match the chromosome
identifiers.}

\item{...}{Additional arguments.}
}
\description{
The \code{DNAStringSet}-based version of the \code{SequenceTrack} class
}
\section{Functions}{
\itemize{
\item \code{initialize(SequenceDNAStringSetTrack)}: Initialize the \code{sequence} slot
(defaulting to an empty \code{\link[Biostrings:DNAStringSet-class]{DNAStringSet}} if
none is supplied) before deferring to the \code{SequenceTrack} initializer for the
remaining slots.

}}
\section{Slots}{

\describe{
\item{\code{sequence}}{A \code{\link[Biostrings:DNAStringSet-class]{DNAStringSet}} object
containing the sequence data, with one element per chromosome.}
}}

