% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceTrack-class.R
\docType{class}
\name{SequenceRNAStringSetTrack-class}
\alias{SequenceRNAStringSetTrack-class}
\alias{initialize,SequenceRNAStringSetTrack-method}
\title{The \code{RNAStringSet}-based version of the \code{SequenceTrack} class}
\usage{
\S4method{initialize}{SequenceRNAStringSetTrack}(.Object, sequence, ...)
}
\arguments{
\item{.Object}{.Object}

\item{sequence}{sequence}

\item{...}{Additional arguments.}
}
\description{
The \code{RNAStringSet}-based version of the \code{SequenceTrack} class
}
\section{Functions}{
\itemize{
\item \code{initialize(SequenceRNAStringSetTrack)}: Initialize the \code{sequence} slot
(defaulting to an empty \code{RNAStringSet} if none is supplied) before
deferring to the \code{SequenceTrack} initializer for the remaining slots.

}}
\section{Slots}{

\describe{
\item{\code{sequence}}{A \code{RNAStringSet} object containing the sequence data, with
one element per chromosome.}
}}

